subscription-backend/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
web3==6.15.1